+++ docs/home/sidebar.ts
@@ -1,1 +1,1 @@
-export const homeSidebar = { text: 'Home', link: '/' }
+export const homeSidebar = { text: 'Home', link: '/' } as const
