+++ docs/learn/conceptual/sidebar.ts
@@ -2,6 +2,7 @@ export const conceptualSidebar = {
   text: '🚧 Everything is an Action',
   collapsed: true,
   items: [
+    { text: '🚧 Why actions', link: '/learn/conceptual/whyactions' },
     { text: '🚧 Actions', link: '/learn/conceptual/actions' },
     { text: '🚧 ActionCreators', link: '/learn/conceptual/actioncreators' },
     { text: '🚧 ActionHandlers', link: '/learn/conceptual/actionhandlers' },
