+++ docs/learn/tutorials/sidebar.ts
@@ -1,0 +1,15 @@
+export const tutorialsSidebar = {
+  text: '🚧 Getting Started',
+  items: [
+    { text: '🚧 Quick start', link: '/learn/gettingstarted/quickstart' },
+    { text: '🚧 Why EVMts', link: '/learn/gettingstarted/why' },
+    {
+      text: '🚧 Tutorial: balanceOf',
+      link: '/learn/tutorials/balanceof',
+    },
+    {
+      text: '🚧 Tutorial: optimistic counter',
+      link: '/learn/tutorials/optimisticcounter',
+    },
+  ],
+} as const
