+++ docs/reference/sidebar.ts
@@ -1,21 +1,11 @@
 import type { DefaultTheme } from 'vitepress'
-import { homeSidebar } from '../common'
-import { actionsReferenceSidebar } from './actions/sidebar'
-import { coreTypesReferenceSidebar } from './coretypes/sidebar'
-import { jsonRpcReferenceSidebar } from './jsonrpc/sidebar'
+import { homeSidebar } from '../home/sidebar'
 
 export const apiReferenceSidebar: DefaultTheme.Sidebar = [
   homeSidebar,
   {
     text: '🚧 API Reference',
     items: [
-      coreTypesReferenceSidebar,
-      actionsReferenceSidebar,
-      jsonRpcReferenceSidebar,
-      buildReferenceSidebar,
-      cliReferenceSidebar,
-      otherUsagesSidebar,
-      advancedSidebar
     ],
   },
 ]
