--- docs/sidebar/common.ts
@@ -1,3 +1,0 @@
-export const homeSidebar = { text: 'Home', link: '/' }
-export const linkToLearnSidebar = { text: 'Learn', link: '/learn/' }
-export const linkToReferenceSidebar = { text: 'Reference', link: '/reference/' }
